結果取得部分切り出し
