+++ app/gui/main_window.py
@@ -8,12 +8,9 @@ import configparser
 import time
 import shutil
 import os
-import sys
 import pyautogui
 import pydirectinput
-from PIL import Image
-from app.lib.vision.vision_api import extract_text_from_image
-from app.lib.opencv.cv2 import compare_images
+from app.lib.crafting_result.crafting_result import get_crafting_result
 from app.lib.macro.macro import consume_prime, skip
 
 
@@ -29,25 +26,6 @@ class MainWindow(tk.Tk):
     """
 
     CONFIG_FILE = "config.ini"
-    TARGET_REGION = (1005, 270, 135, 42)
-
-    @staticmethod
-    def get_resource_path(relative_path):
-        """ ビルドされたEXEファイルに対するリソースの相対パスを取得する """
-        #  pylint: disable=protected-access
-        if hasattr(sys, '_MEIPASS'):
-            return os.path.join(sys._MEIPASS, relative_path)
-        return os.path.join(os.path.abspath("."), relative_path)
-
-    COMPARISON_IMAGES = [
-        get_resource_path.__func__("app/resources/slots/slots441.png"),
-        get_resource_path.__func__("app/resources/slots/slots431.png"),
-        get_resource_path.__func__("app/resources/slots/slots421.png"),
-        get_resource_path.__func__("app/resources/slots/slots411.png"),
-        get_resource_path.__func__("app/resources/slots/slots311.png"),
-        get_resource_path.__func__("app/resources/slots/slots310.png"),
-        get_resource_path.__func__("app/resources/slots/slots300.png")
-    ]
 
     def __init__(self):
         """
@@ -154,8 +132,9 @@ class MainWindow(tk.Tk):
             consume_prime()
             time.sleep(1.65)
             file_name = self.save_screenshot(i + 1)
-            self.perform_ocr(file_name)
-            self.compare_target_images(i + 1)
+            success = get_crafting_result(file_name, i + 1)
+            if success:
+                print("Crafting result obtained successfully")
             time.sleep(0.5)
             skip()
 
@@ -167,7 +146,7 @@ class MainWindow(tk.Tk):
         Args:
             index (int): スクリーンショットのインデックス。
         Returns:
-            None
+            str: 保存したファイル名
         """
         output_dir = "temp"
         output_file_name = f"result_{index}.png"
@@ -182,58 +161,6 @@ class MainWindow(tk.Tk):
         screenshot.save(screenshot_path)
 
         return output_file_name
-
-    def perform_ocr(self, file_name):
-        """
-        tempフォルダ内の画像に対してOCRを実行し、結果をdebug/name.txtに保存するメソッド
-
-        Args:
-            file_name (str): 読み込みファイル名
-        Returns:
-            None
-        """
-        output_dir = "debug"
-        if not os.path.exists(output_dir):
-            os.makedirs(output_dir)
-
-        with open(os.path.join(output_dir, "name.txt"), 'w', encoding='utf-8') as output_file:
-            image_path = os.path.join("temp", file_name)
-            text = extract_text_from_image(image_path)
-            output_file.write(f"{file_name}:\n{text}\n\n")
-
-    def compare_target_images(self, index):
-        """
-        画像を比較し、最も類似度が高い画像を特定するメソッド
-
-        Args:
-            index (int): スクリーンショットのインデックス
-        Returns:
-            None
-        """
-        output_dir = "debug"
-        if not os.path.exists(output_dir):
-            os.makedirs(output_dir)
-
-        # 最初に保存したスクリーンショットのトリミング
-        screenshot_path = os.path.join("temp", f"result_{index}.png")
-        screenshot = Image.open(screenshot_path)
-        x, y, width, height = self.TARGET_REGION
-        target_image = screenshot.crop((x, y, x + width, y + height))
-        target_image_path = os.path.join("temp", f"slots_{index}.png")
-        target_image.save(target_image_path)
-
-        best_match = None
-        best_similarity = 0.0
-
-        for comparison_image in self.COMPARISON_IMAGES:
-            similarity = compare_images(target_image_path, comparison_image)
-            if similarity > best_similarity:
-                best_similarity = similarity
-                best_match = comparison_image
-
-        with open(os.path.join(output_dir, "slots.txt"), 'w', encoding='utf-8') as output_file:
-            output_file.write(
-                f"Best match: {best_match}\nSimilarity: {best_similarity}\n")
 
     def on_closing(self):
         """
